cache_filesystem {memoise}	R Documentation

Filesystem Cache

Description

Use a cache on the local filesystem that will persist between R sessions.

Usage

cache_filesystem(path, algo = "xxhash64", compress = FALSE)

Arguments

path
Directory in which to store cached items.

algo
The hashing algorithm used for the cache, see digest for available algorithms.

compress
Argument passed to saveRDS. One of FALSE, "gzip", "bzip2" or "xz". Default: FALSE.

Examples

## Not run: 
# Use with Dropbox

db <- cache_filesystem("~/Dropbox/.rcache")

mem_runif <- memoise(runif, cache = db)

# Use with Google Drive

gd <- cache_filesystem("~/Google Drive/.rcache")

mem_runif <- memoise(runif, cache = gd)

## End(Not run)

[Package memoise version 2.0.1 Index]