cache_gcs {memoise}	R Documentation

Google Cloud Storage Cache Google Cloud Storage backed cache, for remote caching.

Description

Google Cloud Storage Cache Google Cloud Storage backed cache, for remote caching.

Usage

cache_gcs(
  cache_name = googleCloudStorageR::gcs_get_global_bucket(),
  algo = "sha512",
  compress = FALSE
)

Arguments

cache_name
Bucket name for storing cache files.

algo
The hashing algorithm used for the cache, see digest for available algorithms.

compress
Argument passed to saveRDS. One of FALSE, "gzip", "bzip2" or "xz". Default: FALSE.

Examples

## Not run: 
library(googleCloudStorageR)
# Set GCS credentials.
Sys.setenv("GCS_AUTH_FILE"="<google-service-json>",
           "GCS_DEFAULT_BUCKET"="unique-bucket-name")

gcs <- cache_gcs("unique-bucket-name")
mem_runif <- memoise(runif, cache = gcs)

## End(Not run)

[Package memoise version 2.0.1 Index]