cache_memory {memoise}	R Documentation

In Memory Cache

Description

A cache in memory, that lasts only in the current R session.

Usage

cache_memory(algo = "sha512")

Arguments

algo
The hashing algorithm used for the cache, see digest for available algorithms.

[Package memoise version 2.0.1 Index]