drop_cache {memoise}	R Documentation

Drops the cache of a memoised function for particular arguments.

Description

Drops the cache of a memoised function for particular arguments.

Usage

drop_cache(f)

Arguments

f
Memoised function.

Value

A function, with the same arguments as f, that can be called to drop the cached results of f.

See Also

has_cache, memoise

Examples

mem_sum <- memoise(sum)
mem_sum(1, 2, 3)
mem_sum(2, 3, 4)
has_cache(mem_sum)(1, 2, 3) # TRUE
has_cache(mem_sum)(2, 3, 4) # TRUE
drop_cache(mem_sum)(1, 2, 3) # TRUE
has_cache(mem_sum)(1, 2, 3) # FALSE
has_cache(mem_sum)(2, 3, 4) # TRUE

[Package memoise version 2.0.1 Index]